Move stripe test helper to common setup module
